refactor: extract vaadin-upload-file-list logic to the mixin
